Add functionality to filter supported pages
Added a function to check if the current page is supported or not by the extension. This check is performed using the new function `isSupportedPage()` which compares the URL against a whitelist. The changes were made in response to issues where unsupported pages were throwing errors. This functionality will now bypass those pages and prevent unnecessary requests to the server or unhandled exceptions. URL processing was also encapsulated into `getCurrentURL()` function to improve DRYness.