Refactor URL initialization in content_script.js
Modified the handling of currentTranslatedURL in content_script.js. The initial value of currentTranslatedURL was set to null, and the assignment of a new URL now happens within the startTranslate function. This change ensures that the currentTranslatedURL always corresponds to the currentURL at the time of translation, thus potentially avoiding stale URL issues.