Improve tab interaction and float element functionality
A new method "tabActiveRequestMethod" has been implemented to get the active tab when the tab is activated and then send a message to the active tab if it's on the required URL (developer.apple.com). This improvement enhances the responsiveness of the application as it directly queries the active tab immediately the user switches tabs. If an error occurs, it is logged to the console for debugging.

In addition to this, the script can check if elements exist on the page and remove them if necessary, particularly for the "swiftgg-float" element. The "injectFloat" function now first checks for the existence of "swiftgg-float", preventing unnecessary fetch requests and enhances efficiency.

Finally, the event listeners for the float elements (cancel and translate buttons) have been moved to "content_script.js" for cleaner and more organized code management. Also, fade out animation for the floating elements was added to improve the user interface.