Add CSS classes for 'float' feature
CSS classes have been added for the 'float' feature in content_script.css and float.html file. Previously, in float.html all the styling was written inline and it was harder to manage and troubleshoot. Now, the styling has been moved to the CSS class in content_script.css which makes code cleaner, easier to read and maintain. One of the major advantages of using CSS is that we can create a class once and can use it in multiple places.