Simplify browser detection and update logo source
The browser detection function was refactored to simplify the checks and improve readability. The Safari browser check was moved up to prioritise it. This was done due to Safari's peculiar nature which might contrast with other standard browsers like Chrome and Firefox.

In addition, updated the logo in 'float.html'. The previous URL was lengthy and unsightly, causing clutter in the code. Now changed to a shorter placeholder URL for cleaner code structure.