Add support for whitelisting specific pages
A whitelist feature has been added to popup.js to allow certain pages to bypass translation. This was necessary to prevent unwanted translation of pages that already contain translated content. A whitelist array has been added, and a check has been introduced to see if the active tab's URL is in the whitelist. In background.js, the queryActiveTab function has been updated to reflect these changes. In popup.html, an ID was added to the 'translate' list item for easier DOM manipulation.