"Add category page checking in popup.js"
This commit modifies the condition for enabling our extension on supported pages. It now additionally checks if the current URL is not a category page. This change is necessary to prevent our extension from functioning incorrectly on category pages, by using a new function `isCategoryPage()`, which checks if the last path of the URL is included in the whitelist.