Add 'data-tag' attribute to new elements for easy identification
The update adds a 'data-tag' attribute to all the newly created elements in the DOM, with a value of 'swiftgg'. This is to aid in the identification and manipulation of these elements, without affecting or clashing with existing elements on the page.

Additionally, a new function, 'isInjectedElement', has been added to check if an element has the 'data-tag' attribute and if the value is 'swiftgg', offering a streamlined way of identifying our custom elements.