Simplify plugin state management and update tab handling
We are updating our plugin's state management by removing local storage usage in popup.js file. The unnecessary use of local storage for storing the plugin flag was causing some delay in the update of the interface.

Moreover, some changes have been made to improve the handling of tab updates in the background.js file. Specifically, we added a check to abort further code execution if a tab update event is triggered for the currently active tab that didn't change its URL.

Lastly, we added a new 'removeTranslate' method to the content_script.js file. This method enables cleaning up the translation artifacts from the DOM when the corresponding page’s translation process is cancelled. As a consequence, the commit also includes changes in the background.js file, where we send a message to invoke this method when necessary.

These changes should improve performance and user experience by reducing unnecessary operations and keeping the interface seamlessly updated with the plugin state.