"Improved browser detection and handled CSP rule for Firefox"
This commit comes with several changes directed towards Firefox browser. Firstly, the browser detection function has been updated to more accurately target Firefox. In earlier version, it was only checking the existence of the `browser` object which was not solely unique to Firefox. The update makes it to asynchronously obtain the browser information and compare it with known names.

In addition to that, a new functionality has been introduced in order to disable Content Security Policy (CSP) when the extension is running on Firefox. It operates by adding a new rule to modify specific response headers.

There are also changes noted in `manifest.json` permissions to allow 'declarativeNetRequest' and 'browsingData' for handling this new CSP rule.

Minor changes noted in CSS are done for consistency, replacing 'px' unit with its equivalent number value.