Added display method request and removal of original elements
This commit includes the addition of the displayMethodRequestMethod across popup.js, background.js, and content_script.js. It ensures that users can choose how they want the translated content to be displayed. The logic behind changing the display method was also added to the content_script.js file. A corresponding queryDisplayMethodRequestMethod was also included to get the current display method.

Additionally, an "original" tag was attached to all original elements while translating. A new function, removeOriginal was added to remove all original elements when the display method is set to "chinese". This option was given to help users who prefer viewing only the translated content.

Some code simplification was also carried out, especially with functions that determine the current browser in use. The getUserAgent was removed as it was not used anymore.