Switch from innerText to textContent in translation functions
This commit transitions the "addReplaceTranslate" and "cancelReplaceTranslate"
functions from using .innerText to .textContent. The change was necessary
because .innerText can cause unintended side effects when dealing with style
and layout. It can also lead to performance issues as it forces layout
reflow, while .textContent is faster and more reliable for plain text content.

Signed-off-by: Underthestars-zhy <zhuhaoyu0909@icloud.com>